refactor: extract insight period glance panels
