+++ frontend/src/insights/components/PeriodGlanceCard.tsx
@@ -1,25 +1,18 @@
-import { useEffect, useMemo, useRef, useState, type CSSProperties } from 'react'
+import { useMemo } from 'react'
 import { Sparkles } from 'lucide-react'
 import type { FxStatus } from '@/api/shared/fx'
-import IconTooltip from '@/components/IconTooltip'
 import type {
   PeriodGlancePrimaryMetric,
   PeriodGlanceSupportItem,
-  PeriodGlanceTone,
 } from '@/insights/types/periodGlance'
-import { formatCurrency } from '@/utils/formatCurrency'
-import { getPeriodIncomeExpenseFxStatusMessage } from '@/insights/utils/fxTooltipMessages'
-import { FxStatusBadge } from './FxStatusBadge'
 import {
   InsightLoadingContent,
   InsightLoadingOverlay,
 } from './InsightLoadingTransition'
+import { PeriodGlancePrimaryPanel } from './period-glance/PeriodGlancePrimaryPanel'
+import { PeriodGlanceSupportGrid } from './period-glance/PeriodGlanceSupportGrid'
 import { SectionHeader } from './SectionHeader'
 import { useInsightLoadingSnapshot } from './useInsightLoadingSnapshot'
-
-const PRIMARY_AMOUNT_MAX_REM = 3
-const PRIMARY_AMOUNT_MIN_REM = 1.875
-const PERIOD_GLANCE_TITLE_CONTROL_SLOT_CLASS = 'inline-flex w-[1.375rem] shrink-0 items-center'
 
 type PeriodGlanceSnapshot = {
   primaryMetric: PeriodGlancePrimaryMetric
@@ -41,88 +34,9 @@ type PeriodGlanceCardProps = {
   transitionKey: string
 }
 
-function metricToneClass(tone: PeriodGlanceTone) {
-  if (tone === 'positive') return 'text-[var(--app-positive)]'
-  if (tone === 'negative') return 'text-[var(--app-negative)]'
-  return ''
-}
-
-function MetricCalculationTooltip({ label, calculation }: { label: string, calculation?: string }) {
-  if (!calculation) return null
-
-  return (
-    <IconTooltip
-      label={`${label} calculation`}
-      placement="top"
-      widthClassName="w-72"
-      size={14}
-      strokeWidth={2.25}
-    >
-      {calculation}
-    </IconTooltip>
-  )
-}
-
-function useFittedPrimaryAmount(value: string) {
-  const textRef = useRef<HTMLParagraphElement>(null)
-  const [fontSizeRem, setFontSizeRem] = useState(PRIMARY_AMOUNT_MAX_REM)
-
-  useEffect(() => {
-    const textElement = textRef.current
-    const containerElement = textElement?.parentElement
-    if (!textElement || !containerElement) return undefined
-    const measuredTextElement = textElement
-
-    let frameId = 0
-    let cancelled = false
-
-    function measure() {
-      window.cancelAnimationFrame(frameId)
-      frameId = window.requestAnimationFrame(() => {
-        if (cancelled) return
-
-        const previousFontSize = measuredTextElement.style.fontSize
-        measuredTextElement.style.fontSize = `${PRIMARY_AMOUNT_MAX_REM}rem`
-
-        const availableWidth = measuredTextElement.clientWidth
-        const requiredWidth = measuredTextElement.scrollWidth
-
-        measuredTextElement.style.fontSize = previousFontSize
-
-        const nextFontSize =
-          availableWidth > 0 && requiredWidth > availableWidth
-            ? Math.max(PRIMARY_AMOUNT_MIN_REM, PRIMARY_AMOUNT_MAX_REM * (availableWidth / requiredWidth))
-            : PRIMARY_AMOUNT_MAX_REM
-
-        setFontSizeRem((currentFontSize) =>
-          Math.abs(currentFontSize - nextFontSize) < 0.02 ? currentFontSize : nextFontSize,
-        )
-      })
-    }
-
-    measure()
-    document.fonts?.ready.then(measure)
-
-    if (typeof ResizeObserver === 'undefined') {
-      return () => {
-        cancelled = true
-        window.cancelAnimationFrame(frameId)
-      }
-    }
-
-    const resizeObserver = new ResizeObserver(measure)
-    resizeObserver.observe(containerElement)
-
-    return () => {
-      cancelled = true
-      window.cancelAnimationFrame(frameId)
-      resizeObserver.disconnect()
-    }
-  }, [value])
-
-  return [textRef, fontSizeRem] as const
-}
-
+/**
+ * Renders the period glance summary and supporting insight metrics
+ */
 export function PeriodGlanceCard({
   primaryMetric,
   supportItems,
@@ -151,9 +65,6 @@ export function PeriodGlanceCard({
     loading,
     transitionKey,
   })
-  const [primaryAmountRef, primaryAmountFontSizeRem] = useFittedPrimaryAmount(displaySnapshot.primaryMetric.value)
-  const primaryAmountStyle: CSSProperties | undefined =
-    primaryAmountFontSizeRem < PRIMARY_AMOUNT_MAX_REM ? { fontSize: `${primaryAmountFontSizeRem}rem` } : undefined
 
   return (
     <section className="app-card">
@@ -162,107 +73,14 @@ export function PeriodGlanceCard({
       <div className="relative overflow-hidden" data-tooltip-bounds>
         <InsightLoadingContent concealed={contentConcealed} shouldReduceMotion={shouldReduceMotion}>
           <div className="grid gap-4 min-[1500px]:grid-cols-[minmax(0,40fr)_minmax(0,60fr)]">
-            <div className="grid gap-5 rounded-xl border border-[var(--app-accent-border)] bg-[var(--app-accent-soft)] p-4 min-[750px]:grid-cols-[minmax(0,60fr)_minmax(0,40fr)] min-[750px]:items-center min-[1500px]:flex min-[1500px]:min-h-52 min-[1500px]:flex-col min-[1500px]:items-stretch min-[1500px]:justify-between">
-              <div className="min-w-0 [container-type:inline-size]">
-                <p className="app-label inline-flex items-center gap-2">
-                  {displaySnapshot.primaryMetric.label}
-                  <MetricCalculationTooltip
-                    label={displaySnapshot.primaryMetric.label}
-                    calculation={displaySnapshot.primaryMetric.calculation}
-                  />
-                  {displaySnapshot.incomeExpenseFxStatus && (
-                    <FxStatusBadge
-                      label="Income and expense FX status"
-                      status={displaySnapshot.incomeExpenseFxStatus}
-                      getMessage={getPeriodIncomeExpenseFxStatusMessage}
-                    />
-                  )}
-                </p>
-                <p
-                  ref={primaryAmountRef}
-                  className={[
-                    'mt-3 max-w-full whitespace-nowrap font-financial text-5xl font-normal leading-none',
-                    metricToneClass(displaySnapshot.primaryMetric.tone),
-                  ].join(' ')}
-                  style={primaryAmountStyle}
-                >
-                  {displaySnapshot.primaryMetric.value}
-                </p>
-                <p className="mt-2 max-w-3xl text-sm leading-6 text-[var(--app-text-muted)]">
-                  {displaySnapshot.primaryMetric.detail}
-                </p>
-              </div>
-
-              <div className="grid grid-cols-2 gap-4 border-t border-[var(--app-border)] pt-3 min-[750px]:grid-cols-1 min-[750px]:border-l min-[750px]:border-t-0 min-[750px]:pl-5 min-[750px]:pt-0 min-[1500px]:mt-5 min-[1500px]:grid-cols-2 min-[1500px]:border-l-0 min-[1500px]:border-t min-[1500px]:pl-0 min-[1500px]:pt-3">
-                <div>
-                  <p className="app-label app-label-compact inline-flex items-center gap-2">
-                    Income
-                    <MetricCalculationTooltip
-                      label="Income"
-                      calculation="Total money in for this range after refunds and reversals are netted. Transfers are excluded"
-                    />
-                  </p>
-                  <p className="mt-1 font-financial text-lg">{formatCurrency(displaySnapshot.income, displaySnapshot.displayCurrency)}</p>
-                </div>
-                <div>
-                  <p className="app-label app-label-compact inline-flex items-center gap-2">
-                    Expenses
-                    <MetricCalculationTooltip
-                      label="Expenses"
-                      calculation="Total money out for this range after refunds and reversals are netted. Shown as a positive amount. Transfers are excluded"
-                    />
-                  </p>
-                  <p className="mt-1 font-financial text-lg">{formatCurrency(displaySnapshot.expenses, displaySnapshot.displayCurrency)}</p>
-                </div>
-              </div>
-            </div>
-
-            <div className="grid min-[750px]:grid-cols-[minmax(0,40fr)_minmax(0,60fr)]">
-              {displaySnapshot.supportItems.map((item, index) => (
-                <div
-                  key={item.label}
-                  className={[
-                    'border-[var(--app-border)] py-4 min-[750px]:p-4',
-                    index < displaySnapshot.supportItems.length - 1 ? 'border-b' : '',
-                    index === 0 ? 'pt-0 min-[750px]:border-r min-[750px]:pl-0 min-[750px]:pt-0' : '',
-                    index === 1 ? 'min-[750px]:pr-0 min-[750px]:pt-0' : '',
-                    index === 2
-                      ? 'min-[750px]:border-b-0 min-[750px]:border-r min-[750px]:pb-0 min-[750px]:pl-0'
-                      : '',
-                    index === 3 ? 'pb-0 min-[750px]:pr-0' : '',
-                  ].join(' ')}
-                >
-                  <div className="flex min-h-28 flex-col items-center justify-center text-center">
-                    <p className="app-label inline-flex items-center justify-center gap-2">
-                      <span className={`${PERIOD_GLANCE_TITLE_CONTROL_SLOT_CLASS} justify-end`}>
-                        {item.fxStatus && (
-                          <FxStatusBadge
-                            label={`${item.label} FX status`}
-                            status={item.fxStatus}
-                            getMessage={item.getFxStatusMessage}
-                          />
-                        )}
-                      </span>
-                      {item.label}
-                      <span className={`${PERIOD_GLANCE_TITLE_CONTROL_SLOT_CLASS} justify-start`}>
-                        <MetricCalculationTooltip
-                          label={item.label}
-                          calculation={item.calculation}
-                        />
-                      </span>
-                    </p>
-                    <p
-                      className={['mt-1 text-2xl font-semibold leading-tight', metricToneClass(item.tone)].join(' ')}
-                    >
-                      {item.value}
-                    </p>
-                    <p className="mt-2 text-sm leading-6 text-[var(--app-text-muted)]">
-                      {item.detail}
-                    </p>
-                  </div>
-                </div>
-              ))}
-            </div>
+            <PeriodGlancePrimaryPanel
+              primaryMetric={displaySnapshot.primaryMetric}
+              income={displaySnapshot.income}
+              expenses={displaySnapshot.expenses}
+              incomeExpenseFxStatus={displaySnapshot.incomeExpenseFxStatus}
+              displayCurrency={displaySnapshot.displayCurrency}
+            />
+            <PeriodGlanceSupportGrid supportItems={displaySnapshot.supportItems} />
           </div>
         </InsightLoadingContent>
